+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - New flag for inline mode: `--output/-o` to redirect output to file (will use STDOUT if not set)
 - New `mangal config set` command to set config values. See `mangal help config set` for more info.
 - New `mangal config get` command. See `mangal help config get` for more info 
+- Improve `mangal clear` command. It's more accurate and faster now. 
 
 ## 3.7.0